LogoutContent를 ReadCustomer에 합치기
로그인에도 flexbox 적용
