+++ Armful/jeonha_redux/src/component/LoginRoot.js
@@ -1,7 +1,6 @@
 import { findByLabelText } from '@testing-library/react';
 import React, {Component} from 'react';
 import LoginContent from '../containers/LoginContent';
-import LogoutContent from '../containers/LogoutContent';
 import ReadCustomer from '../containers/ReadCustomer';
 import store from '../store';
 
@@ -16,14 +15,6 @@ export default class LoginRoot extends Component {
         }.bind(this));
     }
 
-    getLogout(){
-        let _article = null;
-        if(this.state.mode === 'readCustomer'){
-            _article = <LogoutContent></LogoutContent>
-        }
-        return _article;
-    }
-
     getLogin(){
         let _article = null;
         if(this.state.mode === 'login'){
@@ -34,16 +25,9 @@ export default class LoginRoot extends Component {
         return _article;
       }
     render() {
-        // const LoginStyle = {
-        //     position: 'fixed',
-        //     top: 90,
-        //     left: 30,
-        //     minWidth: 50
-        //   }
         return (
             <div>
                 {this.getLogin()}
-                {this.getLogout()}
             </div>
         )
     }
