+++ Armful/jeonha_redux/src/component/ReadCustomer.js
@@ -5,7 +5,10 @@ class ReadCustomer extends Component {
 
       return (
           <article>
-              안녕하세요. {this.props.ID}님!
+              안녕하세요. {this.props.ID}님! &nbsp;
+              <input type="button" value="LOGOUT" onClick={function () {
+                this.props.onClick()
+              }.bind(this)}></input>
           </article>
       );
     }
