+++ Armful/jeonha_redux/src/component/Welcome.js
@@ -2,11 +2,6 @@ import React, { Component } from 'react';
 
 class Welcome extends Component {
     render() {
-      // const welcomeStyle = {
-      //   position: 'fixed',
-      //   top: 150,
-      //   left: 30
-      // }
       return (
           <article>
               <h2>환영합니다!</h2>
