+++ Armful/jeonha_redux/src/containers/ReadCustomer.js
@@ -1,10 +1,18 @@
 import ReadCustomer from "../component/ReadCustomer";
 import {connect} from 'react-redux';
 
+function mapDispatchToProps(dispatch){
+    return {
+        onClick:function(){
+            dispatch({type:'LOGOUT',ID:null,PW:null});
+        }
+    }
+}
+
 function mapReduxStateToReactProps(state) {
     return{
         ID:state.ID
     }    
 }
 
-export default connect(mapReduxStateToReactProps)(ReadCustomer);
+export default connect(mapReduxStateToReactProps,mapDispatchToProps)(ReadCustomer);
